IITR HUGGING FACE/requirements.txt
opencv-python
Pillow
gradio
torch
easyocr
numpy
numpy 